+++ Nagisa/XamlResources/CustomConverters.xaml.h
@@ -67,4 +67,59 @@ namespace Nagisa
 			Object ^parameter,
 			String^ language);
 	};
+
+	[Windows::UI::Xaml::Data::Bindable]
+	[Windows::Foundation::Metadata::WebHostHidden]
+	public ref class StatusErrorToVisibleConverter sealed : IValueConverter
+	{
+	public:
+		// Inherited via IValueConverter
+		virtual Object^ Convert(
+			Object^ value,
+			TypeName targetType,
+			Object^ parameter,
+			String^ language);
+		virtual Object^ ConvertBack(
+			Object^ value,
+			TypeName targetType,
+			Object ^parameter,
+			String^ language);
+	};
+
+	[Windows::UI::Xaml::Data::Bindable]
+	[Windows::Foundation::Metadata::WebHostHidden]
+	public ref class StatusPausedToVisibleConverter sealed : IValueConverter
+	{
+	public:
+		// Inherited via IValueConverter
+		virtual Object^ Convert(
+			Object^ value,
+			TypeName targetType,
+			Object^ parameter,
+			String^ language);
+		virtual Object^ ConvertBack(
+			Object^ value,
+			TypeName targetType,
+			Object ^parameter,
+			String^ language);
+	};
+
+	[Windows::UI::Xaml::Data::Bindable]
+	[Windows::Foundation::Metadata::WebHostHidden]
+	public ref class StatusRunningToVisibleConverter sealed : IValueConverter
+	{
+	public:
+		// Inherited via IValueConverter
+		virtual Object^ Convert(
+			Object^ value,
+			TypeName targetType,
+			Object^ parameter,
+			String^ language);
+		virtual Object^ ConvertBack(
+			Object^ value,
+			TypeName targetType,
+			Object ^parameter,
+			String^ language);
+	};
+
 }