+++ Shared/M2AsyncHelpers.h
@@ -11,8 +11,10 @@ License: The MIT License
 #define _M2_ASYNC_HELPERS_
 
 #include <Windows.h>
-#include <wrl\event.h>
+#include <wrl\client.h>
+
 #include "M2RemoveReference.h"
+#include "M2Object.h"
 
 // The M2SetAsyncCompletedHandler function sets the completed handler function 
 // for asynchronous call.
@@ -27,12 +29,66 @@ License: The MIT License
 // Return value:
 //
 // This function does not return a value.
-template<typename TAsync, typename TFunction>
-void M2SetAsyncCompletedHandler(TAsync Async, TFunction Function)
+template<typename TAsync, typename... TFunction>
+inline void M2SetAsyncCompletedHandler(
+	TAsync Async, const TFunction&... Function)
 {
-	Async->Completed =
-		ref new M2RemoveReference<decltype(Async->Completed)>::Type(Function);
+	Async->Completed = 
+		ref new M2RemoveReference<decltype(Async->Completed)>::Type(
+			Function...);
 }
+
+// The M2SetAsyncProgressHandler function sets the progress handler function 
+// for asynchronous call.
+//
+// Parameters:
+//
+// Async
+//     The asynchronous call you want to set.
+// Function
+//     The progress handler function you want to set.
+//
+// Return value:
+//
+// This function does not return a value.
+template<typename TAsync, typename... TFunction>
+inline void M2SetAsyncProgressHandler(
+	TAsync Async, const TFunction&... Function)
+{
+	Async->Progress = 
+		ref new M2RemoveReference<decltype(Async->Progress)>::Type(
+			Function...);
+}
+
+// The M2GetInspectable function retrieves the IInspectable interface from the 
+// provided C++/CX object. 
+//
+// Parameters:
+//
+// object
+//     The C++/CX object you want to retrieve the raw pointer.
+//
+// Return value:
+//
+// The return value is the IInspectable interface from the provided C++/CX 
+// object.
+inline IInspectable* M2GetInspectable(Platform::Object^ object) throw()
+{
+	return reinterpret_cast<IInspectable*>(object);
+}
+
+// The M2AsyncHandleCompleted function handles the completed asynchronous call.
+//
+// Parameters:
+//
+// Async
+//     The completed asynchronous call you want to handle.
+//
+// Return value:
+//
+// The return value is determined by the asynchronous call.
+// The function will return error code if the function fails. 
+HRESULT M2AsyncHandleCompleted(Platform::Object^ Async);
 
 // The M2AsyncWait function uses the non-blocking way to try to wait 
 // asynchronous call.
@@ -51,10 +107,10 @@ void M2SetAsyncCompletedHandler(TAsync Async, TFunction Function)
 // The return value is determined by the asynchronous call.
 // The function will throw a COM exception if the function fails. 
 template<typename TAsync>
-inline auto M2AsyncWait(
+auto M2AsyncWait(
 	TAsync Async, int32 Timeout = -1) -> decltype(Async->GetResults())
 {
-	using Microsoft::WRL::Wrappers::Event;
+	using M2::CHandle;
 	using Platform::COMException;
 	using Windows::Foundation::AsyncStatus;
 	using Windows::Foundation::IAsyncInfo;
@@ -62,44 +118,23 @@ inline auto M2AsyncWait(
 	HRESULT hr = S_OK;
 
 	// Create an event object for wait the asynchronous call.
-	Event CompletedEvent(
+	CHandle CompletedEvent(
 		CreateEventExW(nullptr, nullptr, 0, EVENT_ALL_ACCESS));
-	if (CompletedEvent.IsValid())
+	if (!CompletedEvent.IsInvalid())
 	{
-		// Set event signaled to end wait when completed.
+		// Set the completed handler.
 		M2SetAsyncCompletedHandler(
-			Async, [&](IAsyncInfo^ asyncInfo, AsyncStatus asyncStatus)
+			Async, [&](IAsyncInfo^ /*asyncInfo*/, AsyncStatus /*asyncStatus*/)
 		{
-			UNREFERENCED_PARAMETER(asyncInfo);
-			UNREFERENCED_PARAMETER(asyncStatus);
-
-			SetEvent(CompletedEvent.Get());
+			// Set event signaled to end wait when completed.
+			SetEvent(CompletedEvent);
 		});
 
 		// Wait event object signaled.
-		WaitForSingleObjectEx(CompletedEvent.Get(), Timeout, FALSE);
+		WaitForSingleObjectEx(CompletedEvent, Timeout, FALSE);
 
-		switch (Async->Status)
-		{
-		case AsyncStatus::Completed:
-			// Just do nothing if succeeded.
-			break;
-		case AsyncStatus::Started:
-			// Cancel the asynchronous call and set error code if the 
-			// status is still Started, the timeout interval has been 
-			// elapsed.
-			Async->Cancel();
-			hr = __HRESULT_FROM_WIN32(ERROR_TIMEOUT);
-			break;
-		case AsyncStatus::Canceled:
-			// If the status is Cancelled, set the error code.
-			hr = E_ABORT;
-			break;
-		default:
-			// If the status is other value, get and set the error code.
-			hr = Async->ErrorCode.Value;
-			break;
-		}
+		// Handle the completed asynchronous call.
+		hr = M2AsyncHandleCompleted(Async);
 	}
 	else
 	{
@@ -117,25 +152,19 @@ inline auto M2AsyncWait(
 	return Async->GetResults();
 }
 
-// The M2SetAsyncProgressHandler function sets the progress handler function 
-// for asynchronous call.
+// The M2ExecuteOnUIThread function executes function on the UI thread with 
+// normal priority.
 //
 // Parameters:
 //
-// Async
-//     The asynchronous call you want to set.
-// Function
-//     The progress handler function you want to set.
+// agileCallback
+//     The function you want to execute.
 //
 // Return value:
 //
-// This function does not return a value.
-template<typename TAsync, typename TFunction>
-void M2SetAsyncProgressHandler(TAsync Async, TFunction Function)
-{
-	Async->Progress =
-		ref new M2RemoveReference<decltype(Async->Progress)>::Type(Function);
-}
+// The return value is Windows::Foundation::IAsyncAction^.
+Windows::Foundation::IAsyncAction^ M2ExecuteOnUIThread(
+	Windows::UI::Core::DispatchedHandler^ agileCallback);
 
 // The M2ExecuteOnUIThread function executes function on the UI thread with 
 // normal priority.
@@ -148,16 +177,12 @@ void M2SetAsyncProgressHandler(TAsync Async, TFunction Function)
 // Return value:
 //
 // The return value is Windows::Foundation::IAsyncAction^.
-template<typename TFunction>
-Windows::Foundation::IAsyncAction^ M2ExecuteOnUIThread(TFunction Function)
+template<typename... TFunction>
+inline Windows::Foundation::IAsyncAction^ M2ExecuteOnUIThread(
+	const TFunction&... Function)
 {
-	using Windows::ApplicationModel::Core::CoreApplication;
-	using Windows::UI::Core::DispatchedHandler;
-	using Windows::UI::Core::CoreDispatcherPriority;
-
-	return CoreApplication::MainView->CoreWindow->Dispatcher->RunAsync(
-		CoreDispatcherPriority::Normal,
-		ref new DispatchedHandler(Function));
+	return M2ExecuteOnUIThread(
+		ref new Windows::UI::Core::DispatchedHandler(Function...));
 }
 
 #endif // _M2_ASYNC_HELPERS_