+++ Shared/NagisaVersion.h
@@ -12,7 +12,7 @@ License: The MIT License
 
 #define NAGISA_VER_MAJOR 0
 #define NAGISA_VER_MINOR 1
-#define NAGISA_VER_BUILD 4
+#define NAGISA_VER_BUILD 5
 #define NAGISA_VER_REV 0
 #endif